mumps routine differ
Can be used to quickly check to see which routines are out of date.
do STT^GITDIFF
different routines are outputted to screen
where a routine has been committed to github, but does not exist in the environment, then a warning is outputted
routine versions are saved to /tmp/git/diff- files with a .me extension are envirnment versions, .git are the git versions.  This allows users to copy the code to their desktops and then diff
do PATCH^GITDIFF(rtn) allows you patch an individual routine.
note: the code obviously only compares routines that have committed to the repository.