BOM (Byte Order Mark) fix.
During file integrity checks, the software has been updated to ignore any BOM (Byte Order Mark) encoding at the start of an uploaded file. If a BOM-encoded file is uploaded, the system will issue a warning instead of an error, notifying the user that the file contains BOM encoding.
The boundary checks currently support curl, Python, and the Vue.js UI. However, users might upload files using other languages or tools. If the code fails to detect boundary data in the HTTP request, an error will be returned. Additionally, the system will generate an error if the first column (ID) is not numeric.